src/forma1_django/requirements.txt
Django
djangorestframework
markdown
django-filter
requests
rocketchat-API
djangorestframework-simplejwt
django-utils-six
countryinfo
pycountry
geocoder
django-rest-swagger
django-allauth
django-email-verification
django-ipware
pytz
pyyaml
drf-spectacular~=0.26.5
django-cors-headers
drf-spectacular[sidecar]
tox
dj-stripe
drf_yasg
conduit
django-sslserver
django-debug-toolbar
django-redis
redis
django-redis-cache
django-storages
python-decouple
django-enumchoicefield
